pbs-80--feat: Text Editor 분리
